move data table components to global /components
